chore: raise ValueErrors for duals in IPs
