+++ optlang/tests/abstract_test_cases.py
@@ -26,6 +26,7 @@ import json
 import copy
 import os
 import sympy
+from functools import partial
 
 __test__ = False
 
@@ -696,6 +697,57 @@ class AbstractModelTestCase(unittest.TestCase):
         self.assertEqual(c1.expression - (2 * x + y), 0)
         self.assertEqual(model.optimize(), optlang.interface.OPTIMAL)
         self.assertAlmostEqual(x.primal, 0.5)
+
+    def test_is_integer(self):
+        model = self.model
+        self.assertFalse(model.is_integer)
+
+        model.variables[0].type = "integer"
+        self.assertTrue(model.is_integer)
+
+        model.variables[0].type = "continuous"
+        model.variables[1].type = "binary"
+        self.assertTrue(model.is_integer)
+
+        model.variables[1].type = "continuous"
+        self.assertFalse(model.is_integer)
+
+    def test_integer_variable_dual(self):
+        model = self.interface.Model()
+        x = self.interface.Variable("x", lb=0)
+        y = self.interface.Variable("y", lb=0)
+        c = self.interface.Constraint(x + y, ub=1)
+        model.add(c)
+        model.objective = self.interface.Objective(x)
+
+        model.optimize()
+        self.assertEqual(y.dual, -1)
+
+        x.type = "integer"
+        model.optimize()
+        self.assertRaises(ValueError, partial(getattr, y, "dual"))
+
+        x.type = "continuous"
+        model.optimize()
+        self.assertEqual(y.dual, -1)
+
+    def test_integer_constraint_dual(self):
+        model = self.interface.Model()
+        x = self.interface.Variable("x")
+        c = self.interface.Constraint(x, ub=1)
+        model.add(c)
+        model.objective = self.interface.Objective(x)
+
+        model.optimize()
+        self.assertEqual(c.dual, 1)
+
+        x.type = "integer"
+        model.optimize()
+        self.assertRaises(ValueError, partial(getattr, c, "dual"))
+
+        x.type = "continuous"
+        model.optimize()
+        self.assertEqual(c.dual, 1)
 
 
 @six.add_metaclass(abc.ABCMeta)
